Fixing dataset issue
