+++ src/components/color-picker/color-picker.component.ts
@@ -3,7 +3,7 @@ import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { combineLatest } from 'rxjs/observable/combineLatest';
 import { pairwise, takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs/Subject';
-import { ThemeColor } from '../../services/color/color.service';
+import { ColorPickerColor } from './color-picker-color';
 
 // Values corresponding to stylesheet
 const BUTTON_MARGIN = 8;
@@ -133,64 +133,3 @@ export type ColorPickerInputColors = ColorPickerColor | string;
 export type ColorPickerButtonStyle = 'square' | 'circle';
 export type ColorPickerButtonSize = 'sm' | 'md' | 'lg';
 export type ColorPickerInputMode = 'hex' | 'rgba';
-
-/**
- * Type representing a color, including its descriptive name.
- */
-export class ColorPickerColor {
-
-    /**
-     * Human-readable name of the color.
-     */
-    name: string;
-
-    /**
-     * Hex value of the color, e.g. `#ffffff`.
-     */
-    get hex(): string {
-        return this._originalHexValue ? this._originalHexValue : this._color.toHex();
-    }
-
-    /**
-     * RGBA value of the color, e.g. `rgba(255, 255, 255, 1)`.
-     */
-    get rgba(): string {
-        return this._originalRgbaValue ? this._originalRgbaValue : this._color.toRgba();
-    }
-
-    get r(): number {
-        return parseInt(this._color.getRed());
-    }
-
-    get g(): number {
-        return parseInt(this._color.getGreen());
-    }
-
-    get b(): number {
-        return parseInt(this._color.getBlue());
-    }
-
-    get a(): number {
-        return parseFloat(this._color.getAlpha());
-    }
-
-    private _color: ThemeColor;
-    private _originalHexValue: string;
-    private _originalRgbaValue: string;
-
-    constructor(name: string, value: string, inputMode?: ColorPickerInputMode) {
-        this.name = name;
-        this._color = ThemeColor.parse(value);
-
-        // Preserve the format entered by the user if it's valid
-        if (inputMode === 'hex') {
-            this._originalHexValue = value;
-        } else if (inputMode === 'rgba') {
-            this._originalRgbaValue = value;
-        }
-    }
-
-    toString(): string {
-        return this._color.toRgba();
-    }
-}
