+++ src/components/color-picker/index.ts
@@ -1,2 +1,4 @@
+export * from './color-picker-color';
 export * from './color-picker.component';
 export * from './color-picker.module';
+
